feat(searcher): integrar soporte multiidioma (i18n) en barra de búsqueda y filtros
