+++ src/components/Searcher/Searcher.jsx
@@ -1,4 +1,5 @@
 import React from "react";
+import { useTranslation } from "react-i18next"; // <-- Sumamos el hook de traducciones
 import iconoLupa from "../../assets/icons/lupa.png";
 import iconoPico from "../../assets/icons/diamond_pickaxe.webp";
 import iconoZombie from "../../assets/icons/zombie_head.png";
@@ -9,6 +10,8 @@ const Searcher = ({
   setCategoriaSeleccionada,
   alFiltrarSecundario,
 }) => {
+  const { t } = useTranslation(); // <-- Inicializamos la función t()
+
   return (
     <div className="w-full bg-transparent border-b-8 border-black py-8 mb-12 px-4 shadow-[inset_0px_8px_15px_rgba(0,0,0,0.5)]">
       <div className="max-w-6xl mx-auto flex flex-col gap-4 font-mono">
@@ -18,7 +21,7 @@ const Searcher = ({
             <img src={iconoLupa} alt="Buscar" className="w-6 h-6 m-4" />
             <input
               type="text"
-              placeholder="BUSCAR ... "
+              placeholder={t("search_placeholder")} // Traducido
               onChange={(evento) => alBuscar && alBuscar(evento.target.value)}
               className="flex-grow p-3 font-mono text-lg outline-none w-full text-black placeholder-gray-400 font-bold uppercase"
             />
@@ -36,7 +39,7 @@ const Searcher = ({
                   : "bg-[#333333] text-white border-black hover:bg-[#444444] shadow-[6px_6px_0px_0px_rgba(0,0,0,1)] hover:shadow-[2px_2px_0px_0px_rgba(0,0,0,1)] hover:translate-y-1 hover:translate-x-1"
               }`}
             >
-              ☰ TODO
+              ☰ {t("all")} {/* Traducido */}
             </button>
 
             <button
@@ -49,7 +52,8 @@ const Searcher = ({
                   : "bg-[#333333] text-white border-black hover:bg-[#444444] shadow-[6px_6px_0px_0px_rgba(0,0,0,1)] hover:shadow-[2px_2px_0px_0px_rgba(0,0,0,1)] hover:translate-y-1 hover:translate-x-1"
               }`}
             >
-              <img src={iconoPico} alt="Items" className="w-6 h-6" /> ITEMS
+              <img src={iconoPico} alt="Items" className="w-6 h-6" />{" "}
+              {t("items")}
             </button>
 
             <button
@@ -62,7 +66,8 @@ const Searcher = ({
                   : "bg-[#333333] text-white border-black hover:bg-[#444444] shadow-[6px_6px_0px_0px_rgba(0,0,0,1)] hover:shadow-[2px_2px_0px_0px_rgba(0,0,0,1)] hover:translate-y-1 hover:translate-x-1"
               }`}
             >
-              <img src={iconoZombie} alt="Mobs" className="w-6 h-6" /> MOBS
+              <img src={iconoZombie} alt="Mobs" className="w-6 h-6" />{" "}
+              {t("mobs")}
             </button>
           </div>
         </div>
@@ -78,22 +83,23 @@ const Searcher = ({
                   }
                   className="border-4 border-black p-2 bg-[#333333] text-white font-bold text-sm uppercase shadow-[4px_4px_0px_0px_rgba(0,0,0,1)]"
                 >
-                  <option value="">UTILIDAD: TODAS</option>
-                  <option value="Ataque">ATAQUE</option>
-                  <option value="Alquimia">ALQUIMIA</option>
-                  <option value="Construcción">CONSTRUCCIÓN</option>
-                  <option value="Alimento">ALIMENTO</option>
+                  <option value="">{t("utility_all")}</option>
+                  {/* Los values quedan en español para no romper el filtro, el texto se traduce */}
+                  <option value="Ataque">{t("attack")}</option>
+                  <option value="Alquimia">{t("alchemy")}</option>
+                  <option value="Construcción">{t("construction")}</option>
+                  <option value="Alimento">{t("food")}</option>
                 </select>
 
                 <select
                   onChange={(e) => alFiltrarSecundario("size", e.target.value)}
                   className="border-4 border-black p-2 bg-[#333333] text-white font-bold text-sm uppercase shadow-[4px_4px_0px_0px_rgba(0,0,0,1)]"
                 >
-                  <option value="">TIPO: TODOS</option>
-                  <option value="Herramienta">HERRAMIENTA</option>
-                  <option value="Bloque">BLOQUE</option>
-                  <option value="Consumible">CONSUMIBLE</option>
-                  <option value="Recurso">RECURSO</option>
+                  <option value="">{t("type_all")}</option>
+                  <option value="Herramienta">{t("tool")}</option>
+                  <option value="Bloque">{t("block")}</option>
+                  <option value="Consumible">{t("consumable")}</option>
+                  <option value="Recurso">{t("resource")}</option>
                 </select>
               </>
             )}
@@ -106,22 +112,22 @@ const Searcher = ({
                   }
                   className="border-4 border-black p-2 bg-[#333333] text-white font-bold text-sm uppercase shadow-[4px_4px_0px_0px_rgba(0,0,0,1)]"
                 >
-                  <option value="">COMPORTAMIENTO: TODOS</option>
-                  <option value="Hostil">HOSTIL</option>
-                  <option value="Neutral">NEUTRAL</option>
-                  <option value="Pasivo">PASIVO</option>
-                  <option value="Amigable">AMIGABLE</option>
+                  <option value="">{t("behavior_all")}</option>
+                  <option value="Hostil">{t("hostile")}</option>
+                  <option value="Neutral">{t("neutral")}</option>
+                  <option value="Pasivo">{t("passive")}</option>
+                  <option value="Amigable">{t("friendly")}</option>
                 </select>
 
                 <select
                   onChange={(e) => alFiltrarSecundario("size", e.target.value)}
                   className="border-4 border-black p-2 bg-[#333333] text-white font-bold text-sm uppercase shadow-[4px_4px_0px_0px_rgba(0,0,0,1)]"
                 >
-                  <option value="">TAMAÑO: TODOS</option>
-                  <option value="Pequeño">PEQUEÑO</option>
-                  <option value="Mediano">MEDIANO</option>
-                  <option value="Grande">GRANDE</option>
-                  <option value="Gigante">GIGANTE</option>
+                  <option value="">{t("size_all")}</option>
+                  <option value="Pequeño">{t("small")}</option>
+                  <option value="Mediano">{t("medium")}</option>
+                  <option value="Grande">{t("large")}</option>
+                  <option value="Gigante">{t("giant")}</option>
                 </select>
               </>
             )}
