--- src/i18n.js
@@ -1,111 +1,0 @@
-import i18n from "i18next";
-import { initReactI18next } from "react-i18next";
-
-const savedLanguage = localStorage.getItem("language") || "es";
-
-const resources = {
-  en: {
-    translation: {
-      bestiary: "BESTIARY",
-      home: "HOME",
-      favorites: "FAVORITES",
-      skin_creator: "SKIN CREATOR",
-      footer_text:
-        "Faculty of Informatics - Advanced Web Programming - TP 2 REACT 2026",
-      role_pm: "Scrum Master / PM",
-      role_dev: "Developer",
-      errorText: "Oops! exploded",
-      buttonError: "Respawn in Lobby",
-      inventoryEmpty: "Your inventory is empty",
-      inventoryText:
-        "No mob or item has been added to the favorites collection",
-      favorites: "FAVORITES",
-      only_items: "ITEMS ONLY",
-      only_mobs: "MOBS ONLY",
-      skin_creator: "SKIN CREATOR",
-      empty: "EMPTY",
-      search_database: "Search Items/Mobs",
-      portal_table: "PORTAL TABLE",
-      portal_desc: "Tap an item to navigate the encyclopedia.",
-      go_back: "Go back",
-      go_to: "GO TO",
-      enter: "ENTER",
-      close: "CLOSE",
-      textFavorite: "Add to favorites",
-      inFavorite: "In Favorites",
-      total_saved: "TOTAL SAVED",
-      mobs: "MOBS",
-      items: "ITEMS",
-      generating_terrain: "Generating terrain",
-      item_lost_lava: "That item fell into lava and was lost.",
-      back_to_home: "Back to Home",
-      go_back: "GO BACK",
-      saved_in_favorites: "Saved in favorites",
-      saved: "SAVED",
-      save: "SAVE",
-      description: "DESCRIPTION",
-      statistics: "STATISTICS",
-      behavior: "Behavior",
-      utility: "Utility",
-      size: "Size",
-      type: "Type",
-    },
-  },
-  es: {
-    translation: {
-      bestiary: "BESTIARIO",
-      home: "INICIO",
-      favorites: "FAVORITOS",
-      skin_creator: "CREADOR DE SKINS",
-      footer_text:
-        "Facultad de Informática - Programación Web Avanzada - TP 2 REACT 2026",
-      role_pm: "Scrum Master / PM",
-      role_dev: "Desarrolladora",
-      errorText: "¡Ups! Explotó",
-      buttonError: "Reaparecer en Lobby",
-      inventoryEmpty: "Tu inventario esta vacío",
-      inventoryText:
-        "No se ha agregado ningún mob o item a la colección de favoritos",
-      favorites: "FAVORITOS",
-      only_items: "SOLO ÍTEMS",
-      only_mobs: "SOLO MOBS",
-      skin_creator: "CREA TU SKIN",
-      empty: "VACÍO",
-      search_database: "Explorar Items/Mobs",
-      portal_table: "MESA DE CRAFTEO",
-      portal_desc: "Toca un ítem para navegar por la enciclopedia.",
-      go_back: "Volver atrás",
-      go_to: "IR A",
-      enter: "ENTRAR",
-      close: "CERRAR",
-      textFavorite: "Agregar a favoritos",
-      inFavorite: "En Favoritos",
-      total_saved: "TOTAL GUARDADOS",
-      mobs: "MOBS",
-      items: "ÍTEMS",
-      generating_terrain: "Generando terreno",
-      item_lost_lava: "Ese elemento cayó a la lava y se perdió.",
-      back_to_home: "Volver al Inicio",
-      go_back: "ATRÁS",
-      saved_in_favorites: "Guardado en favoritos",
-      saved: "GUARDADO",
-      save: "GUARDAR",
-      description: "DESCRIPCIÓN",
-      statistics: "ESTADÍSTICAS",
-      behavior: "Comportamiento",
-      utility: "Utilidad",
-      size: "Tamaño",
-      type: "Tipo",
-    },
-  },
-};
-
-i18n.use(initReactI18next).init({
-  resources,
-  lng: savedLanguage,
-  interpolation: {
-    escapeValue: false,
-  },
-});
-
-export default i18n;
