+++ src/locals/en.js
@@ -1,0 +1,63 @@
+export const en = {
+  bestiary: "BESTIARY",
+  home: "HOME",
+  favorites: "FAVORITES",
+  skin_creator: "SKIN CREATOR",
+  footer_text:
+    "Faculty of Informatics - Advanced Web Programming - TP 2 REACT 2026",
+  role_pm: "Scrum Master / PM",
+  role_dev: "Developer",
+  errorText: "Oops! exploded",
+  buttonError: "Respawn in Lobby",
+  inventoryEmpty: "Your inventory is empty",
+  inventoryText: "No mob or item has been added to the favorites collection",
+  only_items: "ITEMS ONLY",
+  only_mobs: "MOBS ONLY",
+  empty: "EMPTY",
+  search_database: "Search Items/Mobs",
+  portal_table: "PORTAL TABLE",
+  portal_desc: "Tap an item to navigate the encyclopedia.",
+  go_back: "GO BACK",
+  go_to: "GO TO",
+  enter: "ENTER",
+  close: "CLOSE",
+  textFavorite: "Add to favorites",
+  inFavorite: "In Favorites",
+  total_saved: "TOTAL SAVED",
+  mobs: "MOBS",
+  items: "ITEMS",
+  generating_terrain: "Generating terrain",
+  item_lost_lava: "That item fell into lava and was lost.",
+  back_to_home: "Back to Home",
+  saved_in_favorites: "Saved in favorites",
+  saved: "SAVED",
+  save: "SAVE",
+  description: "DESCRIPTION",
+  statistics: "STATISTICS",
+  behavior: "Behavior",
+  utility: "Utility",
+  size: "Size",
+  type: "Type",
+  search_placeholder: "SEARCH ...",
+  all: "ALL",
+  utility_all: "UTILITY: ALL",
+  attack: "ATTACK",
+  alchemy: "ALCHEMY",
+  construction: "CONSTRUCTION",
+  food: "FOOD",
+  type_all: "TYPE: ALL",
+  tool: "TOOL",
+  block: "BLOCK",
+  consumable: "CONSUMABLE",
+  resource: "RESOURCE",
+  behavior_all: "BEHAVIOR: ALL",
+  hostile: "HOSTILE",
+  neutral: "NEUTRAL",
+  passive: "PASSIVE",
+  friendly: "FRIENDLY",
+  size_all: "SIZE: ALL",
+  small: "SMALL",
+  medium: "MEDIUM",
+  large: "LARGE",
+  giant: "GIANT",
+};
