+++ src/locals/es.js
@@ -1,0 +1,64 @@
+export const es = {
+  bestiary: "BESTIARIO",
+  home: "INICIO",
+  favorites: "FAVORITOS",
+  skin_creator: "CREA TU SKIN",
+  footer_text:
+    "Facultad de Informática - Programación Web Avanzada - TP 2 REACT 2026",
+  role_pm: "Scrum Master / PM",
+  role_dev: "Desarrolladora",
+  errorText: "¡Ups! Explotó",
+  buttonError: "Reaparecer en Lobby",
+  inventoryEmpty: "Tu inventario esta vacío",
+  inventoryText:
+    "No se ha agregado ningún mob o item a la colección de favoritos",
+  only_items: "SOLO ÍTEMS",
+  only_mobs: "SOLO MOBS",
+  empty: "VACÍO",
+  search_database: "Explorar Items/Mobs",
+  portal_table: "MESA DE CRAFTEO",
+  portal_desc: "Toca un ítem para navegar por la enciclopedia.",
+  go_back: "ATRÁS",
+  go_to: "IR A",
+  enter: "ENTRAR",
+  close: "CERRAR",
+  textFavorite: "Agregar a favoritos",
+  inFavorite: "En Favoritos",
+  total_saved: "TOTAL GUARDADOS",
+  mobs: "MOBS",
+  items: "ÍTEMS",
+  generating_terrain: "Generando terreno",
+  item_lost_lava: "Ese elemento cayó a la lava y se perdió.",
+  back_to_home: "Volver al Inicio",
+  saved_in_favorites: "Guardado en favoritos",
+  saved: "GUARDADO",
+  save: "GUARDAR",
+  description: "DESCRIPCIÓN",
+  statistics: "ESTADÍSTICAS",
+  behavior: "Comportamiento",
+  utility: "Utilidad",
+  size: "Tamaño",
+  type: "Tipo",
+  search_placeholder: "BUSCAR ...",
+  all: "TODO",
+  utility_all: "UTILIDAD: TODAS",
+  attack: "ATAQUE",
+  alchemy: "ALQUIMIA",
+  construction: "CONSTRUCCIÓN",
+  food: "ALIMENTO",
+  type_all: "TIPO: TODOS",
+  tool: "HERRAMIENTA",
+  block: "BLOQUE",
+  consumable: "CONSUMIBLE",
+  resource: "RECURSO",
+  behavior_all: "COMPORTAMIENTO: TODOS",
+  hostile: "HOSTIL",
+  neutral: "NEUTRAL",
+  passive: "PASIVO",
+  friendly: "AMIGABLE",
+  size_all: "TAMAÑO: TODOS",
+  small: "PEQUEÑO",
+  medium: "MEDIANO",
+  large: "GRANDE",
+  giant: "GIGANTE",
+};
