+++ src/locals/i18n.js
@@ -1,0 +1,21 @@
+import i18n from "i18next";
+import { initReactI18next } from "react-i18next";
+
+import { en } from "./en";
+import { es } from "./es";
+
+const savedLanguage = localStorage.getItem("language") || "es";
+
+i18n.use(initReactI18next).init({
+  resources: {
+    en: { translation: en },
+    es: { translation: es },
+  },
+  lng: savedLanguage,
+  fallbackLng: "es",
+  interpolation: {
+    escapeValue: false,
+  },
+});
+
+export default i18n;
