+++ src/main.jsx
@@ -2,7 +2,7 @@ import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import "./index.css";
 import App from "./App.jsx";
-import "./i18n";
+import "./locals/i18n";
 
 createRoot(document.getElementById("root")).render(
   <StrictMode>
